Renaming toothpick-kotlin to ktp
