+++ deps.gradle
@@ -28,7 +28,7 @@ ext.deps = [// Common
             tp_smoothie_support : 'com.github.stephanenicolas.toothpick:smoothie-support:3.0.0',
             tp                  : 'com.github.stephanenicolas.toothpick:toothpick:3.0.0',
             tp_runtime          : 'com.github.stephanenicolas.toothpick:toothpick-runtime:3.0.0',
-            tp_kotlin           : 'com.github.stephanenicolas.toothpick:toothpick-kotlin:3.0.0',
+            ktp                 : 'com.github.stephanenicolas.toothpick:ktp:3.0.0',
             tp_compiler         : 'com.github.stephanenicolas.toothpick:toothpick-compiler:3.0.0',
             tp_testing          : 'com.github.stephanenicolas.toothpick:toothpick-testing:3.0.0',
             tp_testing_junit4   : 'com.github.stephanenicolas.toothpick:toothpick-testing-junit4:3.0.0',
