+++ ktp/gradle.properties
@@ -1,4 +1,4 @@
-POM_ARTIFACT_ID=toothpick-kotlin
-POM_NAME=Toothpick Kotlin Extensions
+POM_ARTIFACT_ID=ktp
+POM_NAME=Kotlin ToothPick
 POM_DESCRIPTION='Kotlin extensions for toothpick'
 POM_PACKAGING='jar'
