+++ ktp/src/main/kotlin/toothpick/ktp/KTPScope.kt
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package toothpick.kotlin
+package toothpick.ktp
 
 import toothpick.Scope
 import toothpick.config.Module
