+++ ktp/src/main/kotlin/toothpick/ktp/binding/BindingExtensions.kt
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package toothpick.kotlin.binding
+package toothpick.ktp.binding
 
 import toothpick.config.Binding
 import toothpick.config.Module
