+++ ktp/src/main/kotlin/toothpick/ktp/delegate/DelegateProvider.kt
@@ -14,9 +14,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package toothpick.kotlin.delegate
+package toothpick.ktp.delegate
 
-import toothpick.kotlin.KTP
+import toothpick.ktp.KTP
 import kotlin.reflect.KProperty
 
 class DelegateProvider<T : Any>(private val clz: Class<T>, private val name: String?, private val injectionType: InjectionType) {
