+++ ktp/src/main/kotlin/toothpick/ktp/delegate/InjectDelegate.kt
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package toothpick.kotlin.delegate
+package toothpick.ktp.delegate
 
 import toothpick.Scope
 import javax.inject.Provider
