+++ ktp/src/main/kotlin/toothpick/ktp/delegate/InjectionType.kt
@@ -14,8 +14,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package toothpick.kotlin.delegate
+package toothpick.ktp.delegate
 
-inline fun <reified T : Any> inject(name: String? = null) = DelegateProvider(T::class.java, name, InjectionType.EAGER)
-inline fun <reified T : Any> lazy(name: String? = null) = DelegateProvider(T::class.java, name, InjectionType.LAZY)
-inline fun <reified T : Any> provider(name: String? = null) = DelegateProvider(T::class.java, name, InjectionType.PROVIDER)
+enum class InjectionType {
+    EAGER, LAZY, PROVIDER
+}
