+++ ktp/src/test/kotlin/toothpick/ktp/delegate/DelegateProviderTest.kt
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package toothpick.kotlin.delegate
+package toothpick.ktp.delegate
 
 import com.nhaarman.mockitokotlin2.mock
 import org.amshove.kluent.shouldBeInstanceOf
