+++ settings.gradle
@@ -1,7 +1,7 @@
+include ':ktp'
 include ':toothpick'
 include ':toothpick-javax-annotations'
 include ':toothpick-compiler'
-include ':toothpick-kotlin'
 include ':toothpick-runtime'
 include ':toothpick-testing'
 include ':toothpick-testing-junit4'
