+++ smoothie-lifecycle-ktp/build.gradle
@@ -34,7 +34,7 @@ android {
 
 dependencies {
   api project(':smoothie-lifecycle')
-  implementation project(':toothpick-kotlin')
+  implementation project(':ktp')
   implementation deps.lifecycle
   implementation deps.androidxv4
   implementation deps.kotlin_runtime
