+++ smoothie-lifecycle-ktp/src/main/kotlin/toothpick/smoothie/lifecycle/LifecycleUtilExtensions.kt
@@ -17,7 +17,7 @@
 package toothpick.smoothie.lifecycle
 
 import androidx.lifecycle.LifecycleOwner
-import toothpick.kotlin.KTPScope
+import toothpick.ktp.KTPScope
 
 /**
  * The scope will be closed automatically during {@code owner}'s onDestroy event.
