+++ toothpick-kotlin-android-sample/src/main/kotlin/com/example/toothpick/kotlin/RxMVPActivity.kt
@@ -9,12 +9,12 @@ import android.widget.TextView
 import butterknife.BindView
 import butterknife.ButterKnife
 import com.example.toothpick.kotlin.deps.RxPresenter
-import toothpick.kotlin.KTP
+import toothpick.ktp.KTP
 import rx.Subscription
 import rx.functions.Action1
 import toothpick.Scope
 import toothpick.Toothpick
-import toothpick.kotlin.delegate.inject
+import toothpick.ktp.delegate.inject
 import toothpick.smoothie.module.SmoothieActivityModule
 
 class RxMVPActivity : Activity() {
